requirements.txt
matplotlib
tk
python-dotenv
tastytrade